Move chapter 13 code sections.
